backend/requirements_deploy.txt
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
PyPDF2==3.0.1
python-docx==1.2.0
lxml==6.0.0

